+++ lib/sucker_punch.rb
@@ -7,16 +7,18 @@ require 'sucker_punch/version'
 require 'logger'
 
 module SuckerPunch
+  RUNNING = Concurrent::AtomicBoolean.new(true)
+
   class << self
-    def exception_handler(&block)
-      @handler = block
+    def exception_handler=(handler)
+      @exception_handler = handler
     end
 
-    def handler
-      @handler || method(:default_handler)
+    def exception_handler
+      @exception_handler || method(:default_exception_handler)
     end
 
-    def default_handler(ex, klass, args)
+    def default_exception_handler(ex, klass, args)
       msg = "Sucker Punch job error for class: '#{klass}' args: #{args}\n"
       msg += "#{ex.class} #{ex}\n"
       msg += "#{ex.backtrace.nil? ? '' : ex.backtrace.join("\n")}"
@@ -37,7 +39,43 @@ module SuckerPunch
       l
     end
 
+    def shutdown_handler
+      @shutdown_handler || method(:default_shutdown_handler)
+    end
+
+    def shutdown_handler=(handler)
+      @shutdown_handler = handler
+    end
+
+    def default_shutdown_handler
+      if SuckerPunch::RUNNING.make_false
+        stopped = false
+        queues = SuckerPunch::Queue.all
+
+        logger.info("Shutdown triggered...executing remaining in-process jobs")
+
+        stopping = []
+        queues.each do |queue|
+          queue.pool.shutdown
+          stopping << queue
+        end
+
+        stopped = stopping.all? { |queue| queue.pool.wait_for_termination(1) }
+
+        return if stopped
+
+        if !stopped
+          logger.info("Remaining jobs didn't finish in time...killing remaining jobs")
+          queues.each { |queue| queue.pool.kill }
+        end
+      end
+    end
   end
+end
+
+at_exit do
+  SuckerPunch.shutdown_handler.call
+  SuckerPunch.logger.info("All is quiet...byebye")
 end
 
 require 'sucker_punch/railtie' if defined?(::Rails)