+++ lib/sucker_punch/job.rb
@@ -30,14 +30,16 @@ module SuckerPunch
 
     module ClassMethods
       def perform_async(*args)
+        return unless SuckerPunch::RUNNING.true?
         queue = SuckerPunch::Queue.find_or_create(self.to_s, num_workers)
-        queue.post(args) { |args| __run_perform(*args) }
+        queue.pool.post(args) { |args| __run_perform(*args) }
       end
 
       def perform_in(interval, *args)
+        return unless SuckerPunch::RUNNING.true?
         queue = SuckerPunch::Queue.find_or_create(self.to_s, num_workers)
-        job = Concurrent::ScheduledTask.execute(interval.to_f, args: args, executor: queue) do |args|
-          self.new.perform(*args)
+        job = Concurrent::ScheduledTask.execute(interval.to_f, args: args, executor: queue.pool) do
+          __run_perform(*args)
         end
         job.pending?
       end
@@ -53,7 +55,7 @@ module SuckerPunch
         result
       rescue => ex
         SuckerPunch::Counter::Failed.new(self.to_s).increment
-        SuckerPunch.handler.call(ex, self, args)
+        SuckerPunch.exception_handler.call(ex, self, args)
       ensure
         SuckerPunch::Counter::Busy.new(self.to_s).decrement
       end